ANDROID: sdcardfs: Don't iput if we didn't igrab
If we fail to get top, top is either NULL, or igrab found
that we're in the process of freeing that inode, and did
not grab it. Either way, we didn't grab it, and have no
business putting it.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 38117720
Change-Id: Ie2f587483b9abb5144263156a443e89bc69b767b
Signed-off-by: Kevin F. Haggerty <haggertk@lineageos.org>
(cherry picked from commit 3f5c14de5e082349b7033bb7186d2ba3d79ba778)